fix reporte incidentes por empleado

+++ servicio-reportes/src/main/java/servicioreportes/repositories/IncidenteRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import servicioreportes.entities.Incidente;
 import java.util.List;
 
-public interface IncidenteRepository
-        extends JpaRepository<Incidente, Long> {
-    List<Incidente> findByEmpleadoId(Long empleadoId);
+public interface IncidenteRepository extends JpaRepository<Incidente, Long> {
+    List<Incidente> findByEmpleadoId(int empleadoId);
 }
